% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dbox}
\alias{dbox}
\title{King County ballot drop box locations}
\format{
A data frame with columns including \code{lat} and \code{long}.
}
\source{
King County Elections
}
\usage{
dbox
}
\description{
A data frame of ballot drop box locations in King County, Washington.
}
\keyword{datasets}
