% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{early_meck}
\alias{early_meck}
\title{Mecklenburg County early voting locations}
\format{
A data frame with columns including \code{lat} and \code{long}.
}
\source{
Mecklenburg County Board of Elections
}
\usage{
early_meck
}
\description{
A data frame of early voting locations in Mecklenburg County,
North Carolina.
}
\keyword{datasets}
