% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{king_geo}
\alias{king_geo}
\title{King County voter sample with geocoded addresses}
\format{
A data frame with columns including \code{Residence_Addresses_Latitude}
and \code{Residence_Addresses_Longitude}.
}
\source{
King County voter file (anonymized sample)
}
\usage{
king_geo
}
\description{
A sample of geocoded voter records from King County, Washington,
including latitude and longitude of residential addresses.
}
\keyword{datasets}
