% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rvoterdistance-package.R
\docType{package}
\name{Rvoterdistance-package}
\alias{Rvoterdistance}
\alias{Rvoterdistance-package}
\title{Rvoterdistance: Voter Distance to Polling Locations}
\description{
Calculates the distance between each voter in a voter file (given lat/long coordinates or sf point geometries) and multiple polling or vote-by-mail drop box locations. Returns nearest location, k-nearest locations, or all locations within a distance threshold. Core computation uses the Haversine formula implemented in C++ via 'Rcpp'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lorenc5/Rvoterdistance}
  \item Report bugs at \url{https://github.com/lorenc5/Rvoterdistance/issues}
}

}
\author{
\strong{Maintainer}: Loren Collingwood \email{lcollingwood@unm.edu}

}
\keyword{internal}
