\name{Rvoterdistance-package}
\alias{Rvoterdistance-package}
\alias{Rvoterdistance}
\docType{package}
\title{
  Calculates the Distance Between Voter and Multiple Polling Locations
}
\description{
  This package is designed to calculate the distance between each voter in a voter file -- given lat/long coordinates -- and many potential (early) polling or vote by mail drop box locations, then return the minimum distance.
}
\details{
See demo(demo, "Rvoterdistance") for examples on how to use the code}
\author{
Loren Collingwood, UC Riverside

Maintainer: Loren Collingwood <loren.collingwood@ucr.edu>
}
%\references{
%  This optional section can contain literature or other references %for background information.
%}
%\keyword{ package }
%\seealso{
%  Optional links to other man pages
%}
%\examples{
%  \dontrun{
%     ## Optional simple examples of the most important functions
%     ## These can be in \dontrun{} and \donttest{} blocks.   
%  }
%}
